+++ vicinus/__init__.py
@@ -6,6 +6,6 @@ from vicinus.ngf_idf import (
     ngf_idf_setup, ngf_idf_query, ngf_idf_rank, ngf_idf_select,
 )
 from vicinus.ngrams import n_grams, jaccard_index
+from vicinus.samples import list_samples, get_sample
 from vicinus.sparse_vector import SparseVector
 from vicinus.tokenize import tokenize
-